# zoo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/animal": {
            "get": {
                "description": "Возвращает информацию о животном по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Получение информации о животном",
                "parameters": [
                    {
                        "description": "ID животного",
                        "name": "animal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.GetAnimalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.GetAnimalResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Животное не найдено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет новое животное в вольер",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Создание нового животного",
                "parameters": [
                    {
                        "description": "Новое животное",
                        "name": "animal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.AddAnimalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID нового животного",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет животное из вольера",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Удаление животного",
                "parameters": [
                    {
                        "description": "ID животного",
                        "name": "animal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.DeleteAnimalRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешно удалено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Животное не найдено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/animal/move": {
            "post": {
                "description": "Перемещает животное в другой вольер",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Перемещение животного в другой вольер",
                "parameters": [
                    {
                        "description": "ID животного и ID нового вольера",
                        "name": "animal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ChangeEnclosureRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно перемещено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Животное или вольер не найдены",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/animals": {
            "get": {
                "description": "Возвращает список всех животных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Получение всех животных",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.GetAllAnimalsResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "animal.Type": {
            "type": "string",
            "enum": [
                "predator",
                "herbivore",
                "bird",
                "aquarium"
            ],
            "x-enum-varnames": [
                "Predator",
                "Herbivore",
                "Bird",
                "Aquarium"
            ]
        },
...
        "http.AddAnimalRequest": {
            "type": "object",
            "properties": {
                "animal_type": {
                    "$ref": "#/definitions/animal.Type"
                },
                "enclosure_id": {
                    "type": "integer"
                },
                "favorite_food": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "http.ChangeEnclosureRequest": {
            "type": "object",
            "properties": {
                "animal_id": {
                    "type": "integer"
                },
                "enclosure_id": {
                    "type": "integer"
                }
            }
        },
...
        "http.DeleteAnimalRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "http.GetAllAnimalsResponse": {
            "type": "object",
            "properties": {
                "animals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.GetAnimalResponse"
                    }
                }
            }
        },
...
        "http.GetAnimalRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "http.GetAnimalResponse": {
            "type": "object",
            "properties": {
                "animal_type": {
                    "$ref": "#/definitions/animal.Type"
                },
                "enclosure_id": {
                    "type": "integer"
                },
                "favorite_food": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}